### 2020-08-30 25fa71b
added some attributes to simplify your_header
