### 2020-08-30 e1ab26b
function to check if file exists
