### 2020-08-30 5ea1cef
check if file exists and is of correct type
