### 2020-08-30 ac46ee5
simplifying handling of different formats
