+++ your/your.py
@@ -7,9 +7,13 @@ import numpy as np
 
 from your.formats.psrfits import PsrfitsFile
 from your.formats.pysigproc import SigprocFile
+from your.utils.misc import check_file_exist, MyEncoder
 
 logger = logging.getLogger(__name__)
 
+formats = {}
+formats['fil'] = SigprocFile
+formats['fits'] = PsrfitsFile
 
 class Your(PsrfitsFile, SigprocFile):
     """
@@ -39,43 +43,38 @@ class Your(PsrfitsFile, SigprocFile):
         self.your_file = file
         if isinstance(self.your_file, str):
             ext = os.path.splitext(self.your_file)[1]
+            check_file_exist(self.your_file)
             if ext == ".fits" or ext == ".sf":
-                logger.debug(f'Reading the fits file: {self.your_file}')
-                PsrfitsFile.__init__(self, psrfitslist=[self.your_file])
-                self.isfits = True
-                self.isfil = False
+                self.format = 'fits'
+                self.your_file = [self.your_file]
             elif ext == ".fil":
-                logger.debug(f'Reading filterbank file {self.your_file}')
-                SigprocFile.__init__(self, fp=self.your_file)
-                self.isfits = False
-                self.isfil = True
+                self.format = 'fil'
             else:
-                raise TypeError('Filetype not supported')
+                raise TypeError(f'Filetype {ext} not supported')
         elif isinstance(self.your_file, list):
             if len(self.your_file) == 0:
                 raise ValueError('Filelist is empty. Please check the input')
             if len(self.your_file) == 1 and os.path.splitext(*self.your_file)[1] == ".fil":
-                for filterbank_file in self.your_file:
-                    logger.debug(f'Reading filterbank file {filterbank_file}')
-                    SigprocFile.__init__(self, fp=filterbank_file)
-                    self.isfits = False
-                    self.isfil = True
+                self.format = 'fil'
+                self.your_file = self.your_file[0]
+                check_file_exist(self.your_file)
             else:
                 for f in self.your_file:
+                    check_file_exist(f)
                     ext = os.path.splitext(f)[1]
                     if ext == ".fits" or ext == ".sf":
                         pass
                     else:
                         raise TypeError("Can only work with a list of fits file or an individual filterbank file")
                 self.your_file.sort()
-                logger.debug(f'Reading the following fits files: {self.your_file}')
-                PsrfitsFile.__init__(self, psrfitslist=self.your_file)
-                self.isfits = True
-                self.isfil = False
+                self.format = 'fits'
         else:
             raise ValueError(
                 'file should be a string of input file path or a list of strings with relevant file paths.')
 
+        logger.debug(f'Reading the file(s): {self.your_file}')
+        self.formatclass = formats[self.format]
+        self.formatclass.__init__(self, self.your_file)
         if not self.source_name:
             logger.info(f'Source name not present in the file. Setting source name to TEMP')
             self.source_name = 'TEMP'
@@ -97,10 +96,7 @@ class Your(PsrfitsFile, SigprocFile):
         Returns: Native sampling time of the data in seconds
 
         """
-        if self.isfil:
-            return SigprocFile.native_tsamp(self)
-        else:
-            return PsrfitsFile.native_tsamp(self)
+        return self.formatclass.native_tsamp(self)
 
     @property
     def native_foff(self):
@@ -109,10 +105,7 @@ class Your(PsrfitsFile, SigprocFile):
         Returns: Native channel bandwidth of the data in MHz
 
         """
-        if self.isfil:
[…]
