### 2020-08-30 a2c8d4e
added format to your_header. getting rid of isfits and isfil
