+++ paper/paper.md
@@ -66,6 +66,8 @@ candidates in real-time or near-real-time. Typically the data output from the te
 [psrfits](https://github.com/scottransom/presto/blob/master/python/presto/psrfits.py), 
 [sigpyproc](https://github.com/FRBs/sigpyproc3), etc). Still, no general tool exists which can work across data formats.
 
+
+# Statement of need 
 `Your` (Your Unified Reader) is a python-based library that unifies the data processing across multiple commonly 
 used formats. `Your` was conceived initially to perform data ingestion for The Petabyte FRB search Project (TPP), which 
 will uniformly search a large number of datasets from telescopes around the world for FRBs. As this 
