+++ paper/paper.bib
@@ -75,7 +75,7 @@ archivePrefix = {arXiv},
       volume={17},
       pages={261--272},
       adsurl = {https://rdcu.be/b08Wh},
-      doi = {https://doi.org/10.1038/s41592-019-0686-2},
+      doi = {10.1038/s41592-019-0686-2},
 }
 
 @software{reback2020pandas,
@@ -150,7 +150,8 @@ Bdsk-Url-1 = {https://doi.org/10.3847/1538-3881/aabc4f}}
   number={2},
   pages={22},
   year={2011},
-  publisher={IEEE Computer Society}
+  publisher={IEEE Computer Society},
+  doi={10.1109/mcse.2011.37}
 }
 
 @ARTICLE{hotan2004,
