### 2020-09-09 647abf9
can read multiple fits files too
