### 2020-09-11 4ed4dae
fixed logic of nsamps in to_fil, added assert
