### 2020-09-18 716eaa8
updated calculation of data_step for dada, needs testing
