+++ your/__init__.py
@@ -1,4 +1,4 @@
 from your.writer import *
 from your.your import *
 
-__version__ = "0.5.1"
+__version__ = "0.5.2"
