### 2020-09-22 1a06cf5
fix for time length warning
