### 2020-09-22 d1f3752
fixing logic for nsamp_read in get_chunk
