### 2020-09-25 c5ea1fe
Allow custom display resolution
Signed-off-by: Devansh Agarwal <devansh.kv@gmail.com>
