+++ utils/heimdall/heimdall.md
@@ -1,3 +1,5 @@
+<span style="float:right;">[[source]](https://github.com/thepetabyteproject/your/blob/master/your/utils/heimdall.py#L50)</span>
+
 ### HeimdallManager
 
 
@@ -73,6 +75,8 @@ Args:
 
 
 ----
+
+<span style="float:right;">[[source]](https://github.com/thepetabyteproject/your/blob/master/your/utils/heimdall.py#L127)</span>
 
 ### run
 
