+++ your/your.md
@@ -1,3 +1,5 @@
+<span style="float:right;">[[source]](https://github.com/thepetabyteproject/your/blob/master/your/your.py#L14)</span>
+
 ### Your
 
 
@@ -29,6 +31,8 @@ Attributes:
 
 ----
 
+<span style="float:right;">[[source]](https://github.com/thepetabyteproject/your/blob/master/your/your.py#L151)</span>
+
 ### bandpass
 
 
@@ -50,6 +54,8 @@ Returns:
 
 
 ----
+
+<span style="float:right;">[[source]](https://github.com/thepetabyteproject/your/blob/master/your/your.py#L178)</span>
 
 ### get_data
 
@@ -84,6 +90,8 @@ Returns:
 
 ----
 
+<span style="float:right;">[[source]](https://github.com/thepetabyteproject/your/blob/master/your/your.py#L267)</span>
+
 ### dispersion_delay
 
 
@@ -104,6 +112,8 @@ Returns:
 
 
 ----
+
+<span style="float:right;">[[source]](https://github.com/thepetabyteproject/your/blob/master/your/your.py#L284)</span>
 
 ### Header
 
