+++ cli/cli.md
@@ -9,12 +9,12 @@ Command Line Interface
 
 |Script|Description|
 | :--- | :--- |
-|[your_h5plotter.py](bin/your_h5plotter.md)|Plot candidate h5 files|
-|[your_combine_mocks.py](bin/your_combine_mocks.md)|Combine two bands from mock spectrometer to a filterbank file.|
-|[your_bandpass.py](bin/your_bandpass.md)|Generate bandpass of data|
-|[your_header.py](bin/your_header.md)|Read header from fits/fil files and print the your header|
-|[your_rfimask.py](bin/your_rfimask.md)|Make Bad channel mask|
-|[your_heimdall.py](bin/your_heimdall.md)|Your Heimdall Fetch FRB|
-|[your_writer.py](bin/your_writer.md)|Convert/Write files from any format to a single file in any format.|
-|[your_viewer.py](bin/your_viewer.md)|Read fits/fil file and show the data|
 |[your_candmaker.py](bin/your_candmaker.md)|Your candmaker! Make h5 candidates from the candidate csv files|
+|[your_header.py](bin/your_header.md)|Read header from fits/fil files and print the your header|
+|[your_viewer.py](bin/your_viewer.md)|Read fits/fil file and show the data|
+|[your_writer.py](bin/your_writer.md)|Convert/Write files from any format to a single file in any format.|
+|[your_heimdall.py](bin/your_heimdall.md)|Your Heimdall Fetch FRB|
+|[your_combine_mocks.py](bin/your_combine_mocks.md)|Combine two bands from mock spectrometer to a filterbank file.|
+|[your_h5plotter.py](bin/your_h5plotter.md)|Plot candidate h5 files|
+|[your_rfimask.py](bin/your_rfimask.md)|Make Bad channel mask|
+|[your_bandpass.py](bin/your_bandpass.md)|Generate bandpass of data|

[…]
