### 2020-09-25 3984536
Fixed the Contributing Guide location
Signed-off-by: Devansh Agarwal <devansh.kv@gmail.com>
