### 2020-09-25 66399a1
calc_N does not need nchans
Signed-off-by: Devansh Agarwal <devansh.kv@gmail.com>
