+++ your/utils/rfi.py
@@ -1,8 +1,9 @@
+import logging
 
 import numpy as np
 from scipy import stats
 from scipy.signal import savgol_filter as sg
-import logging
+
 logger = logging.getLogger(__name__)
 
 
@@ -11,17 +12,12 @@ def savgol_filter(bandpass, channel_bandwidth, frequency_window=15, sigma=6):
     Apply savgol filter to the data. See [Agarwal el al. 2020](https://arxiv.org/abs/2003.14272) for details.
 
     Args:
-    
         bandpass (numpy.ndarray): bandpass of the data
-
         channel_bandwidth (float): channel bandwidth (MHz)
-
         frequency_window (float): frequency window (MHz)
-        
         sigma (float): sigma value to apply cutoff on
 
     Returns:
-
         numpy.ndarray: mask for channels
 
     """
@@ -38,15 +34,11 @@ def spectral_kurtosis(data, N=1, d=None):
     Compute spectral kurtosis. See [Nita et al. (2016)](https://doi.org/10.1109/RFINT.2016.7833535) for details.
 
     Args:
-
         data (numpy.ndarray): 2D frequency time data
-
         N (int): Number of accumulations on the FPGA
-
         d (float): shape factor
 
-
-
+    Returns:
          numpy.ndarray: Spectral Kurtosis along frequency axis
 
     """
@@ -64,24 +56,15 @@ def sk_filter(data, channel_bandwidth, nchans, tsamp, N=None, d=1, sigma=5):
     Apply Spectral Kurtosis filter to the data
 
     Args:
-
         data (numpy.ndarray): 2D frequency time data
-
         channel_bandwidth (float): channel bandwidth (MHz)
-
-        nchans (int): number of channels 
-
+        nchans (int): number of channels
         tsamp (float): sampling time (seconds)
-        
         N (int): Number of accumulations on the FPGA
-
         d (float): shape factor
-
         sigma (float): sigma value to apply cutoff on
 
-
     Returns:
-
          numpy.ndarray: mask for channels
 
     """
@@ -98,27 +81,21 @@ def sk_filter(data, channel_bandwidth, nchans, tsamp, N=None, d=1, sigma=5):
     return ~mask
 
 
-def calc_N(channel_bandwidth, nchans, tsamp):
+def calc_N(channel_bandwidth, tsamp):
     """
 
     Calculates number of accumulations on FPGA
 
     Args:
-
         channel_bandwidth (float): channel bandwidth (MHz)
-
-        nchans (int): number of channels
-
         tsamp (float): sampling time (seconds)
 
-
     Returns:
-
         int: FPGA accumulation length
 
     """
 
-    tn = nchans * np.abs(1 / (2 * channel_bandwidth * nchans * 10 ** 6))
+    tn = np.abs(1 / (2 * channel_bandwidth * 10 ** 6))
     return np.round(tsamp / tn)
 
 
@@ -128,22 +105,15 @@ def sk_sg_filter(data, your_object, nchans, spectral_kurtosis_sigma=6, savgol_fr
     Apply Spectral Kurtosis and Savgol filter to the data
 
     Args:
-
         data (numpy.ndarray): 2D frequency time data
-
         your_object: Your object
-
         nchans (int): number of channels
-
         spectral_kurtosis_sigma (float): sigma value to apply cutoff on for SK filter
-
         savgol_frequency_window (float): frequency window for savgol filter(MHz)
-
         savgol_sigma (float): sigma value to apply cutoff on for savgol filter
 
 
     Returns:
-
          numpy.ndarray: mask for channels
 
     """
