### 2020-09-27 4a75f08
Added type hints to almost all functions
Signed-off-by: Devansh Agarwal <devansh.kv@gmail.com>
