+++ your/formats/psrfits.py
@@ -31,12 +31,10 @@ def unpack_2bit(data):
     Unpack 2-bit data that has been read in as bytes.
 
     Args:
-
-        data: array of unsigned 2-bit ints packed into an array of bytes.
+        data (np.ndarray) : array of unsigned 2-bit ints packed into an array of bytes.
 
     Returns:
-
-        unpacked array. The size of this array will be four times the size of the input data.
+        np.ndarray: unpacked array. The size of this array will be four times the size of the input data.
 
     """
     piece0 = np.bitwise_and(data >> 0x06, 0x03)
@@ -51,12 +49,10 @@ def unpack_4bit(data):
     Unpack 4-bit data that has been read in as bytes.
 
     Args:
-
-        data: array of unsigned 4-bit ints packed into an array of bytes.
+        data (np.ndarray) : array of unsigned 4-bit ints packed into an array of bytes.
 
     Returns:
-
-        unpacked array. The size of this array will be twice the size of the input data.
+        np.ndarray: unpacked array. The size of this array will be twice the size of the input data.
 
     """
     piece0 = np.bitwise_and(data >> 0x04, 0x0F)
@@ -72,55 +68,31 @@ class PsrfitsFile(object):
     [psrfits](https://github.com/scottransom/presto/blob/master/python/presto/psrfits.py ).
 
     Args:
-
         psrfitslist (str): list of files
 
     Attributes:
-
         filename (str): Name of the first file
-
         filelist (list): List of files
-        
         fileid (int): Index of the current file
-        
         fits (obj): fits object of the current file read
-        
         specinfo (obj): Object of class SpectraInfo for the given file list
-        
         header (list): Header of the fits file
-        
         source_name (str): Source Name
-
         machine_id (int) : Machine ID
-
         barycentric (int): If 1 the data is barycentered
-
         pulsarcentric (int): Is the data in pulsar's frame of reference?
-
         src_raj (float): RA of the source (HHMMSS.SS)
-
         src_deg (float): Dec of the source (DDMMSS.SS)
-
         az_start (float): Telescope Azimuth (degrees)
-
         za_start (float): Telescope Zenith Angle (degrees)
-
         fch1 (float): Frequency of first channel (MHz))
-
         foff (float): Channel bandwidth (MHz)
-
         nchans (int): Number of channels
-
         nbeams (int): Number of beams in the rcvr.
-
         ibeam (int): Beam number
-
         nbits (int): Number of bits the data are recorded in.
-
         tstart (float): Start MJD of the data
-
         tsamp (float): Sampling interval (seconds)
-
         nifs (int): Number of IFs in the data.
         
     """
@@ -167,8 +139,7 @@ class PsrfitsFile(object):
         """
 
         Returns:
-
-            Total number of spectra in all files in filelist
+            int : Total number of spectra in all files in filelist
 
         """
         return int(self.specinfo.spectra_per_subint * np.sum(self.specinfo.num_subint))
@@ -178,8 +149,7 @@ class PsrfitsFile(object):
         Native number of total spectra in all the files. This will be made a property so that it can't be overwritten
 
         Returns:
-
-            Total number of spectra in all files in filelist
+            int : Total number of spectra in all files in filelist
 
         """
         return int(self.specinfo.spectra_per_subint * np.sum(self.specinfo.num_subint))
@@ -189,8 +159,7 @@ class PsrfitsFile(object):
         This will be made a property so that it can't be overwritten.
 
         Returns:
-
-            Native sampling time of the file.
+            float : Native sampling time of the file.
 
         """
         return self.specinfo.dt
@@ -200,8 +169,7 @@ class PsrfitsFile(object):
         This will be made a property so that it can't be overwritten.
 
[…]
