### 2020-09-27 c7f31ca
Proper Notes rendering for mkdocs
Signed-off-by: Devansh Agarwal <devansh.kv@gmail.com>
