+++ index.md
@@ -39,9 +39,8 @@ pip install git+https://github.com/thepetabyteproject/your.git
 
 
 
-**Note**:
-To use `psrdada`, you would need to install [psrdada-python](https://github.com/AA-ALERT/psrdada-python). 
-[`your_heimdall.py`](https://thepetabyteproject.github.io/your/bin/your_heimdall/) requires [Heimdall](https://sourceforge.net/projects/heimdall-astro/) and [psrdada-python](https://github.com/AA-ALERT/psrdada-python). 
+!!! note
+    To use the `psrdada` format, you would need to install [psrdada-python](https://github.com/AA-ALERT/psrdada-python). [`your_heimdall.py`](https://thepetabyteproject.github.io/your/bin/your_heimdall/) requires [Heimdall](https://sourceforge.net/projects/heimdall-astro/) and [psrdada-python](https://github.com/AA-ALERT/psrdada-python). 
 
 
 # Documentation
