### 2020-09-27 8915a67
fixed calc_N test case
Signed-off-by: Devansh Agarwal <devansh.kv@gmail.com>
