+++ your/formats/dada.py
@@ -2,11 +2,7 @@ import logging
 import os
 
 import numpy as np
-from astropy.time import Time
 from psrdada import Writer
-from tqdm import tqdm
-
-from your.utils.math import primes, closest_divisor, find_gcd
 
 logger = logging.getLogger(__name__)
 
@@ -20,12 +16,15 @@ class DadaManager:
     A manager class for `psrdada` writer.
 
     Args:
+
         size (int): size of each buffer (in bytes)
+
         key (hex): hexadecimal dada key
+
         n_readers (int): Number of dada readers.
     """
 
-    def __init__(self, size, key=0xdada, n_readers=1):
+    def __init__(self, size, key=0xDADA, n_readers=1):
         self.size = size
         self.key = key
         self.n_readers = n_readers
@@ -35,9 +34,13 @@ class DadaManager:
         Kill any previous buffers with the same key.
         Set up the dada buffers and connect to a writer.
         """
-        logger.debug(f"Destroying previous buffers using: dada_db -d -k {self.key} 2>/dev/null")
+        logger.debug(
+            f"Destroying previous buffers using: dada_db -d -k {self.key} 2>/dev/null"
+        )
         os.system(f"dada_db -d -k {self.key} 2>/dev/null")
-        logger.info(f"Creating new buffers using dada_db -b {self.size} -k {self.key} -r {self.n_readers}")
+        logger.info(
+            f"Creating new buffers using dada_db -b {self.size} -k {self.key} -r {self.n_readers}"
+        )
         os.system(f"dada_db -b {self.size} -k {self.key} -r {self.n_readers}")
         self.writer = Writer()
         self.writer.connect(int(self.key, 16))
@@ -53,13 +56,14 @@ class DadaManager:
         Dump the data to the buffer
 
         Args:
+
             data_input (numpy.ndarray): Numpy array of the data.
 
         """
         page = self.writer.getNextPage()
         data = np.asarray(page)
         data.fill(0)
-        data[:len(data_input)] = data_input
+        data[: len(data_input)] = data_input
 
     def mark_filled(self):
         """
@@ -80,105 +84,3 @@ class DadaManager:
         """
         self.writer.disconnect()
         os.system(f"dada_db -d -k {self.key} 2> /dev/null")
-
-
-class YourDada:
-    """
-
-    Linker class between `psrdada` and `your`.
-
-    Args:
-        your_object: your object
-
-    """
-
-    def __init__(self, your_object):
-
-        self.your_object = your_object
-        if self.your_object.your_header.format == 'fits':
-            logger.debug(f'Calculating dada size and data step for the fits files')
-            self.list_of_subints = self.your_object.specinfo.num_subint.astype('int')
-            if len(self.list_of_subints) > 1:
-                # if there is more than one files see how many subints we need to read so that the data is equally split
-                self.subint_steps = int(find_gcd(self.list_of_subints))
-            else:
-                # if there is just one large file, read it in chunks
-                self.subint_steps = int(np.max(np.prod(np.unique((primes(self.list_of_subints))))))
-            self.dada_size = self.subint_steps * self.your_object.your_header.nchans * self.your_object.specinfo.spectra_per_subint  # * self.your_object.nbits / 8  # bytes
-            self.data_step = int(self.subint_steps * self.your_object.specinfo.spectra_per_subint)
-        else:
-            nsamp_gulp = 2 ** 18
-            logger.debug(f'Calculating dada size and data step for the filterbank file')
-            if self.your_object.your_header.nspectra < nsamp_gulp:
-                self.dada_size = self.your_object.your_header.nspectra * self.your_object.your_header.nchans * self.your_object.your_header.nbits / 8  # bytes
[…]

[…]
