+++ bin/your_writer/your_writer.md
@@ -10,11 +10,7 @@ Convert/Write files from any format to a single file in any format.
 
 
 ```bash
-usage: your_writer.py [-h] [-v] -f FILES [FILES ...] -t TYPE [-c CHANS CHANS]
-                      [-nstart NSTART] [-nsamp NSAMP] [-o OUTDIR]
-                      [-name OUT_NAME] [--no_progress] [-r]
-                      [-sksig SPECTRAL_KURTOSIS_SIGMA] [-sgsig SAVGOL_SIGMA]
-                      [-sgfw SAVGOL_FREQUENCY_WINDOW] [-zero_dm_subt]
+usage: your_writer.py [-h] [-v] -f FILES [FILES ...] -t TYPE [-c CHANS CHANS] [-nstart NSTART] [-nsamp NSAMP] [-o OUTDIR] [-name OUT_NAME] [--no_progress] [-r] [-sksig SPECTRAL_KURTOSIS_SIGMA] [-sgsig SAVGOL_SIGMA] [-sgfw SAVGOL_FREQUENCY_WINDOW] [-zero_dm_subt]
 
 ```
 # Arguments
