### 2020-09-30 c4b775a
Documentation bug fixes
Signed-off-by: Devansh Agarwal <devansh.kv@gmail.com>
