### 2020-10-01 1a0e8ec
Bump nbconvert from 5.6.1 to 6.0.6
Bumps [nbconvert](https://github.com/jupyter/nbconvert) from 5.6.1 to 6.0.6.
- [Release notes](https://github.com/jupyter/nbconvert/releases)
- [Commits](https://github.com/jupyter/nbconvert/compare/5.6.1...6.0.6)

Signed-off-by: dependabot[bot] <support@github.com>
