+++ docs/requirements.txt
@@ -3,4 +3,4 @@ mkdocs>=1.1.2
 mkdocs-material>=6.0
 argmark>=0.1
 mkdocs-jupyter>=0.13.0
-nbconvert==5.6.1
+nbconvert==6.0.6
