+++ your/formats/filwriter.py
@@ -1,109 +1,151 @@
 import logging
-import os
-
-import numpy as np
 
 from your.formats.pysigproc import SigprocFile
 
 logger = logging.getLogger(__name__)
 
 
-def make_sigproc_obj(filfile, your_object, nchans, chan_freq, nstart):
+def sigproc_object_from_writer(your_writer):
     """
-    Use Your class to make Sigproc class object with relevant parameters
+
+    Convert a `your_writer` object to Sigproc object for writing
 
     Args:
-        filfile (str) : Name of the Filterbank file
-        your_object : Your object for the PSRFITS files
-        nchans (int) : No. of channels in the frequency range
-        chan_freq (np.ndarray) : An array of required frequency channel range
-        nstart (int): Start sample
+        your_writer: `your_writer` object
 
     Returns:
-        Object of class SigprocFile
+        sigproc object
 
     """
-    logger.debug(f'Generating Sigproc object')
+    logger.debug(f"Generating Sigproc object")
     fil_obj = SigprocFile()
+    fil_obj.rawdatafile = your_writer.outname
+    fil_obj.source_name = your_writer.your_object.your_header.source_name
 
-    logger.debug(f'Setting attributes of Sigproc object from Your object.')
-    fil_obj.rawdatafile = filfile
-    fil_obj.source_name = your_object.your_header.source_name
-
-    # Verify the following parameters
     fil_obj.machine_id = 0  # use "Fake" for now
     fil_obj.barycentric = 0  # by default the data isn't barycentered
     fil_obj.pulsarcentric = 0
     fil_obj.telescope_id = 6  # use only GBT for now
     fil_obj.data_type = 0
 
-    fil_obj.nchans = nchans
-    fil_obj.foff = your_object.your_header.foff
-    fil_obj.fch1 = chan_freq[0]
+    fil_obj.nchans = your_writer.nchans
+    fil_obj.foff = your_writer.your_object.your_header.foff
+    fil_obj.fch1 = your_writer.chan_freqs[0]
     fil_obj.nbeams = 1
     fil_obj.ibeam = 0
-    fil_obj.nbits = your_object.your_header.nbits
-    fil_obj.tsamp = your_object.your_header.tsamp
-    if not nstart:
-        nstart = 0
-    fil_obj.tstart = your_object.your_header.tstart + nstart * your_object.your_header.tsamp / (60 * 60 * 24)
+    fil_obj.nbits = your_writer.your_object.your_header.nbits
+    fil_obj.tsamp = your_writer.your_object.your_header.tsamp
+
+    fil_obj.tstart = your_writer.tstart
     fil_obj.nifs = 1  # Only use Intensity values
 
-    if your_object.your_header.ra_deg and your_object.your_header.dec_deg:
-        ra = your_object.your_header.ra_deg
-        dec = your_object.your_header.dec_deg
+    if (
+            your_writer.your_object.your_header.ra_deg
+            and your_writer.your_object.your_header.dec_deg
+    ):
+        ra = your_writer.your_object.your_header.ra_deg
+        dec = your_writer.your_object.your_header.dec_deg
     else:
         ra = 0
         dec = 0
 
     from astropy.coordinates import SkyCoord
-    loc = SkyCoord(ra, dec, unit='deg')
+    import numpy as np
+
+    loc = SkyCoord(ra, dec, unit="deg")
     ra_hms = loc.ra.hms
     dec_dms = loc.dec.dms
 
-    fil_obj.src_raj = float(f'{int(ra_hms[0]):02d}{np.abs(int(ra_hms[1])):02d}{np.abs(ra_hms[2]):07.4f}')
[…]
