### 2020-10-06 972caa7
Added gulp size and cleaner filterbank writing
Signed-off-by: Devansh Agarwal <devansh.kv@gmail.com>
