+++ your/writer.py
@@ -8,7 +8,7 @@ import tqdm
 from astropy.io import fits
 from astropy.time import Time
 
-from your.formats.filwriter import write_fil
+from your.formats.filwriter import sigproc_object_from_writer
 from your.formats.fitswriter import initialize_psrfits
 from your.utils.math import primes
 from your.utils.rfi import sk_sg_filter
@@ -34,13 +34,14 @@ class Writer:
         spectral_kurtosis_sigma (float) : Sigma for spectral kurtosis filter
         savgol_frequency_window (float) : Filter window for savgol filter
         savgol_sigma (float) : Sigma for savgol filter
+        gulp (int) : Gulp size for the data
         zero_dm_subt (bool) : Enable zero DM rfi excision
 
     """
 
     def __init__(self, your_object, nstart=0, nsamp=None, c_min=None, c_max=None, outdir=None, outname=None,
-                 flag_rfi=False, progress=None, spectral_kurtosis_sigma=4, savgol_frequency_window=15, savgol_sigma=4,
-                 zero_dm_subt=False):
+                 flag_rfi=False, progress=True, spectral_kurtosis_sigma=4, savgol_frequency_window=15, savgol_sigma=4,
+                 gulp=None, zero_dm_subt=False):
 
         self.your_object = your_object
         self.nstart = nstart
@@ -62,6 +63,16 @@ class Writer:
         self.zero_dm_subt = zero_dm_subt
         self.data = None
         self.dada_is_set = False
+
+        if gulp is None:
+            self.gulp = gulp
+        else:
+            p = np.sort(primes(self.nsamp))[::-1]
+            cumprods = np.cumprod(p)
+            self.gulp = int(cumprods[len(cumprods) // 2])
+
+        if self.gulp > self.nsamp:
+            self.gulp = self.nsamp
 
         original_dir, orig_basename = os.path.split(self.your_object.your_header.filename)
         if not self.outname:
@@ -144,37 +155,36 @@ class Writer:
 
         """
 
-        # Calculate loop of spectra
-        if not self.nsamp:
-            self.nsamp = self.your_object.your_header.native_nspectra
+        self.outname += '.fil'
+        with tqdm.tqdm(total=self.nsamp, disable=~self.progress) as pbar:
+            # create the header
+            sigproc_object = sigproc_object_from_writer(self)
 
-        interval = 4096 * 24
-        if self.nsamp < interval:
-            interval = self.nsamp
+            # write the header
+            sigproc_object.write_header(filename=self.outname)
 
-        if self.nsamp > interval:
-            nloops = 1 + self.nsamp // interval
-        else:
-            nloops = 1
-        nstarts = np.arange(self.nstart, self.nstart + interval * nloops, interval, dtype=int)
-        nsamps = np.full(nloops, interval)
-        if nsamps % interval != 0:
-            nsamps = np.append(nsamps, [nsamps % interval])
+            # make sure header got written
+            if not os.path.isfile(self.outname):
+                raise IOError("Failed to write the filterbank file")
 
-        logging.debug(f'nstarts is {nstarts}, nsamps is {nsamps}, nloops is {nloops}, interval is {interval}, '
-                      f'nstart is {self.nstart}')
-        # Read data
-        for st, samp in tqdm.tqdm(zip(nstarts, nsamps), total=len(nstarts), disable=self.progress):
-            logger.debug(f'Reading spectra {st}-{st + samp} in file {self.your_object.your_header.filename}')
-            self.get_data_to_write(st, samp)
-            logger.info(
-                f'Writing data from spectra {st}-{st + samp} in the frequency channel range {self.chan_min}-{self.chan_max} '
-                f'to filterbank')
-            write_fil(self.data, self.your_object, nchans=self.nchans, chan_freq=self.chan_freqs, outdir=self.outdir,
[…]
