### 2020-10-06 5c34d01
fix the gulp size bug
Signed-off-by: Devansh Agarwal <devansh.kv@gmail.com>
