### 2020-10-06 9e62688
if nsamp is prime number then read all of it.
Signed-off-by: Devansh Agarwal <devansh.kv@gmail.com>
