### 2020-10-06 22a8f9a
Fixed the .isfil bug
Signed-off-by: Devansh Agarwal <devansh.kv@gmail.com>
