### 2020-10-14 eaef054
Use writer for RFI mititgation
Signed-off-by: Devansh Agarwal <devansh.kv@gmail.com>
