+++ bin/your_heimdall.py
@@ -12,54 +12,149 @@ import numpy as np
 from your import Your
 from your.utils.heimdall import HeimdallManager
 from your.utils.misc import MyEncoder
-from your.utils.plotter import save_bandpass
-from your.utils.rfi import savgol_filter
 from your.writer import Writer
 
 if __name__ == "__main__":
-    parser = argparse.ArgumentParser(prog='your_heimdall.py', description="Your Heimdall Fetch FRB",
-                                     formatter_class=argparse.ArgumentDefaultsHelpFormatter)
-    parser.add_argument('-v', '--verbose', help='Be verbose', action='count', default=0)
-    parser.add_argument('-p', '--probability', help='Detection threshold', default=0.5, type=float)
-    parser.add_argument('-f', '--files', help='filterbank or psrfits', nargs='+')
-    parser.add_argument('-dm', '--dm', help='DM (eg -dm 10 1000)', type=float, nargs=2, default=[10, 1000])
-    parser.add_argument('-g', '--gpu_id', help='GPU ID to run heimdall on', type=int, required=False, default=0)
-    parser.add_argument('-sg', '--apply_savgol', help='Apply savgol filter to zap bad channels', action='store_true')
-    parser.add_argument('-frequency_window', '--filter_window', help='Window size (MHz) for savgol filter',
-                        required=False,
-                        default=15, type=float)
-    parser.add_argument('-sigma', '--sigma', help='Sigma for the savgol filter', required=False, default=6, type=float)
-    parser.add_argument('-m', '--mask', help='Input RFI mask (could be 1-D bad channel mask or 2-D FT mask)',
-                        required=False, type=str,
-                        default=None)
-    parser.add_argument('-dm_tol', '--dm_tol', help='SNR loss tolerance between DM trials', required=False,
-                        type=float, default=1.25)
-    parser.add_argument('-rfi_no_narrow', '--rfi_no_narrow', help='disable narrow band RFI excision', required=False,
-                        action='store_true', default=False)
-    parser.add_argument('-rfi_no_broad', '--rfi_no_broad', help='disable 0-DM RFI excision', required=False,
-                        action='store_true', default=False)
-    parser.add_argument('-o', '--output_dir', help='Output dir for heimdall candidates', type=str, required=False,
-                        default=None)
-    parser.add_argument('-fs', '--channel_start', help='Start channel for sub band search', type=int, required=False,
-                        default=0)
-    parser.add_argument('-fe', '--channel_end', help='End channel for sub band search', type=int, required=False,
-                        default=-1)
-    parser.add_argument('--no_progress', help='Do not show the tqdm bar', action='store_true', default=None)
+    parser = argparse.ArgumentParser(
+        prog="your_heimdall.py",
+        description="Your Heimdall Fetch FRB",
+        formatter_class=argparse.ArgumentDefaultsHelpFormatter,
+    )
+    parser.add_argument("-v", "--verbose", help="Be verbose", action="count", default=0)
+    parser.add_argument("-f", "--files", help="filterbank or psrfits", nargs="+")
+    parser.add_argument(
+        "-dm",
+        "--dm",
+        help="DM (eg -dm 10 1000)",
+        type=float,
+        nargs=2,
+        default=[10, 1000],
+    )
+    parser.add_argument(
+        "-g",
+        "--gpu_id",
+        help="GPU ID to run heimdall on",
+        type=int,
+        required=False,
+        default=0,
+    )
+    parser.add_argument(
[…]
