### 2020-10-14 663870a
Repaced tqdm with rich
Signed-off-by: Devansh Agarwal <devansh.kv@gmail.com>
