+++ requirements.txt
@@ -5,6 +5,5 @@ scikit-image>=0.14.2
 scipy>=1.3.0
 numba>=0.48.0
 astropy>=4.0
-tqdm>=4.46.0
 pandas>=1.0.3
 rich>=8.0
