### 2020-10-14 13233cf
Richer Logging + option to not write out the logs to files
Signed-off-by: Devansh Agarwal <devansh.kv@gmail.com>
