+++ bin/your_bandpass.py
@@ -29,22 +29,45 @@ def bandpass(f, time=0, nspectra=8192, outname=None):
     save_bandpass(y, bandpass, chan_nos=None, mask=None, outdir=None, outname=outname)
 
 
-if __name__ == '__main__':
-    parser = argparse.ArgumentParser(prog='your_bandpass.py',
-                                     description="Generate bandpass of data",
-                                     formatter_class=argparse.ArgumentDefaultsHelpFormatter)
-    parser.add_argument('-f', '--files',
-                        help='Fits or filterbank files to read.',
-                        required=True, nargs='+')
-    parser.add_argument('-n', '--nspectra',
-                        help='Number of spectra to use for bandpass.',
-                        required=False, type=int, default=8192)
-    parser.add_argument('-t', '--time',
-                        help='Time (s) to use for bandpass.',
-                        required=False, type=float, default=0)
-    parser.add_argument('-b', '--name',
-                        help='Name of bandpass png', type=str,
-                        required=False, default=None)
+if __name__ == "__main__":
+    parser = argparse.ArgumentParser(
+        prog="your_bandpass.py",
+        description="Generate bandpass of data",
+        formatter_class=argparse.ArgumentDefaultsHelpFormatter,
+    )
+    parser.add_argument(
+        "-f",
+        "--files",
+        help="Fits or filterbank files to read.",
+        required=True,
+        nargs="+",
+    )
+    parser.add_argument(
+        "-n",
+        "--nspectra",
+        help="Number of spectra to use for bandpass.",
+        required=False,
+        type=int,
+        default=8192,
+    )
+    parser.add_argument(
+        "-t",
+        "--time",
+        help="Time (s) to use for bandpass.",
+        required=False,
+        type=float,
+        default=0,
+    )
+    parser.add_argument(
+        "-b",
+        "--name",
+        help="Name of bandpass png",
+        type=str,
+        required=False,
+        default=None,
+    )
     values = parser.parse_args()
 
-    bandpass(f=values.files, time=values.time, nspectra=values.nspectra, outname=values.name)
+    bandpass(
+        f=values.files, time=values.time, nspectra=values.nspectra, outname=values.name
+    )
