+++ your/writer.py
@@ -4,9 +4,9 @@ import logging
 import os
 
 import numpy as np
-import tqdm
 from astropy.io import fits
 from astropy.time import Time
+from rich.progress import Progress
 
 from your.formats.filwriter import sigproc_object_from_writer
 from your.formats.fitswriter import initialize_psrfits
@@ -29,7 +29,7 @@ class Writer:
         c_max (int) : End channel index (default: total number of frequencies)
         outdir (str) : Output directory for file
         outname (str) : Name of the file to write to (without the file extension)
-        progress: Turn on/off progress bar
+        progress (bool) : Set to it to false to disable progress bars
         flag_rfi (bool) : To turn on RFI flagging
         spectral_kurtosis_sigma (float) : Sigma for spectral kurtosis filter
         savgol_frequency_window (float) : Filter window for savgol filter
@@ -39,9 +39,23 @@ class Writer:
 
     """
 
-    def __init__(self, your_object, nstart=0, nsamp=None, c_min=None, c_max=None, outdir=None, outname=None,
-                 flag_rfi=False, progress=True, spectral_kurtosis_sigma=4, savgol_frequency_window=15, savgol_sigma=4,
-                 gulp=None, zero_dm_subt=False):
+    def __init__(
+            self,
+            your_object,
+            nstart=0,
+            nsamp=None,
+            c_min=None,
+            c_max=None,
+            outdir=None,
+            outname=None,
+            flag_rfi=False,
+            progress=True,
+            spectral_kurtosis_sigma=4,
+            savgol_frequency_window=15,
+            savgol_sigma=4,
+            gulp=None,
+            zero_dm_subt=False,
+    ):
 
         self.your_object = your_object
         self.nstart = nstart
@@ -78,17 +92,19 @@ class Writer:
         if self.gulp > self.nsamp:
             self.gulp = self.nsamp
 
-        original_dir, orig_basename = os.path.split(self.your_object.your_header.filename)
+        original_dir, orig_basename = os.path.split(
+            self.your_object.your_header.filename
+        )
         if not self.outname:
             name, ext = os.path.splitext(orig_basename)
-            if ext == '.fits':
-                temp = name.split('_')
+            if ext == ".fits":
+                temp = name.split("_")
                 if len(temp) > 1:
-                    self.outname = '_'.join(temp[:-1]) + '_converted'
+                    self.outname = "_".join(temp[:-1]) + "_converted"
                 else:
-                    self.outname = name + '_converted'
+                    self.outname = name + "_converted"
             else:
-                self.outname = name + '_converted'
+                self.outname = name + "_converted"
 
         if not self.outdir:
             self.outdir = original_dir
@@ -113,7 +129,7 @@ class Writer:
 
     @property
     def chan_freqs(self):
-        return self.your_object.chan_freqs[self.chan_min:self.chan_max]
+        return self.your_object.chan_freqs[self.chan_min: self.chan_max]
 
     @property
     def nchans(self):
@@ -121,7 +137,10 @@ class Writer:
 
     @property
     def tstart(self):
-        return self.your_object.your_header.tstart + self.nstart * self.your_object.your_header.tsamp / (60 * 60 * 24)
+        return (
+                self.your_object.your_header.tstart
+                + self.nstart * self.your_object.your_header.tsamp / (60 * 60 * 24)
+        )
 
     def get_data_to_write(self, start_sample, nsamp):
         """
@@ -136,9 +155,16 @@ class Writer:
 
         """
         data = self.your_object.get_data(start_sample, nsamp)
-        data = data[:, self.chan_min:self.chan_max]
+        data = data[:, self.chan_min: self.chan_max]
         if self.flag_rfi:
[…]
