### 2020-10-15 f7515f7
Rich progress bars instead of tqdm
Signed-off-by: Devansh Agarwal <devansh.kv@gmail.com>
