+++ your/formats/fitswriter.py
@@ -47,17 +47,16 @@ class ObsInfo(object):
         self.chan_bw = 0.0
         self.nsblk = 0
 
-        self.telescope = 'VLA'
+        self.telescope = "VLA"
         self.ant_x = -1601185.63
         self.ant_y = -5041978.15
         self.ant_z = 3554876.43
         self.longitude = self.calc_longitude()
 
     def calc_longitude(self):
-        cc = coord.EarthLocation.from_geocentric(self.ant_x,
-                                                 self.ant_y,
-                                                 self.ant_z,
-                                                 unit='m')
+        cc = coord.EarthLocation.from_geocentric(
+            self.ant_x, self.ant_y, self.ant_z, unit="m"
+        )
         longitude = cc.lon.deg
         return longitude
 
@@ -68,7 +67,7 @@ class ObsInfo(object):
         self.stt_imjd = stt_imjd
         self.stt_smjd = stt_smjd
         self.stt_offs = stt_offs
-        self.obs_date = self.format_date(time.Time(mjd, format='mjd').isot)
+        self.obs_date = self.format_date(time.Time(mjd, format="mjd").isot)
 
     def fill_freq_info(self, fcenter, nchan, chan_bw):
         self.fcenter = fcenter
@@ -109,85 +108,150 @@ class ObsInfo(object):
 
     def format_date(self, date_str):
         # Strip out the decimal seconds
-        out_str = date_str.split('.')[0]
+        out_str = date_str.split(".")[0]
         return out_str
 
     def fill_primary_header(self):
         p_hdr = fits.Header()
-        p_hdr["HDRVER"] = ('3.4             ', "Header version                               ")
-        p_hdr["FITSTYPE"] = ('PSRFITS', "FITS definition for pulsar data files        ")
-        p_hdr["DATE"] = (self.file_date, "File creation date (YYYY-MM-DDThh:mm:ss UTC) ")
-        p_hdr["OBSERVER"] = (self.observer, "Observer name(s)                             ")
-        p_hdr["PROJID"] = (self.proj_id, "Project name                                 ")
-        p_hdr["TELESCOP"] = (self.telescope, "Telescope name                               ")
+        p_hdr["HDRVER"] = (
+            "3.4             ",
+            "Header version                               ",
+        )
+        p_hdr["FITSTYPE"] = ("PSRFITS", "FITS definition for pulsar data files        ")
+        p_hdr["DATE"] = (
+            self.file_date,
+            "File creation date (YYYY-MM-DDThh:mm:ss UTC) ",
+        )
+        p_hdr["OBSERVER"] = (
+            self.observer,
+            "Observer name(s)                             ",
+        )
+        p_hdr["PROJID"] = (
+            self.proj_id,
+            "Project name                                 ",
+        )
+        p_hdr["TELESCOP"] = (
+            self.telescope,
+            "Telescope name                               ",
+        )
         p_hdr["ANT_X"] = (self.ant_x, "[m] Antenna ITRF X-coordinate (D)            ")
         p_hdr["ANT_Y"] = (self.ant_y, "[m] Antenna ITRF Y-coordinate (D)            ")
[…]
