### 2020-10-15 bd19227
Removed unwanted chars
Signed-off-by: Devansh Agarwal <devansh.kv@gmail.com>
