+++ your/utils/plotter.py
@@ -3,7 +3,7 @@ import os
 
 import matplotlib
 
-matplotlib.use('Agg')
+matplotlib.use("Agg")
 
 import h5py
 import numpy as np
@@ -25,7 +25,9 @@ def figsize(scale, width_by_height_ratio):
     Returns:
         list: list of width and height
     """
-    fig_width_pt = 513.17  # 469.755                  # Get this from LaTeX using \the\textwidth
+    fig_width_pt = (
+        513.17  # 469.755                  # Get this from LaTeX using \the\textwidth
+    )
     inches_per_pt = 1.0 / 72.27  # Convert pt to inch
     golden_mean = (np.sqrt(5.0) - 1.0) / 2.0  # Aesthetic ratio (you could change this)
     fig_width = fig_width_pt * inches_per_pt * scale  # width in inches
@@ -45,34 +47,46 @@ def get_params(scale=0.5, width_by_height_ratio=1):
     Returns:
         dict: dictionary of parameters
     """
-    params = {'backend': 'pdf',
-              'axes.labelsize': 10,
-              'lines.markersize': 4,
-              'font.size': 10,
-              'xtick.major.size': 6,
-              'xtick.minor.size': 3,
-              'ytick.major.size': 6,
-              'ytick.minor.size': 3,
-              'xtick.major.width': 0.5,
-              'ytick.major.width': 0.5,
-              'xtick.minor.width': 0.5,
-              'ytick.minor.width': 0.5,
-              'lines.markeredgewidth': 1,
-              'axes.linewidth': 1.2,
-              'legend.fontsize': 7,
-              'xtick.labelsize': 10,
-              'ytick.labelsize': 10,
-              'savefig.dpi': 200,
-              'path.simplify': True,
-              'font.family': 'serif',
-              'font.serif': 'Times',
-              'text.latex.preamble': [r'\usepackage{amsmath}', r'\usepackage{amsbsy}',
-                                      r'\DeclareMathAlphabet{\mathcal}{OMS}{cmsy}{m}{n}'],
-              'figure.figsize': figsize(scale, width_by_height_ratio)}
+    params = {
+        "backend": "pdf",
+        "axes.labelsize": 10,
+        "lines.markersize": 4,
+        "font.size": 10,
+        "xtick.major.size": 6,
+        "xtick.minor.size": 3,
+        "ytick.major.size": 6,
+        "ytick.minor.size": 3,
+        "xtick.major.width": 0.5,
+        "ytick.major.width": 0.5,
+        "xtick.minor.width": 0.5,
+        "ytick.minor.width": 0.5,
+        "lines.markeredgewidth": 1,
+        "axes.linewidth": 1.2,
+        "legend.fontsize": 7,
+        "xtick.labelsize": 10,
+        "ytick.labelsize": 10,
+        "savefig.dpi": 200,
+        "path.simplify": True,
+        "font.family": "serif",
+        "font.serif": "Times",
+        "text.latex.preamble": [
+            r"\usepackage{amsmath}",
+            r"\usepackage{amsbsy}",
+            r"\DeclareMathAlphabet{\mathcal}{OMS}{cmsy}{m}{n}",
+        ],
+        "figure.figsize": figsize(scale, width_by_height_ratio),
+    }
     return params
 
 
-def plot_h5(h5_file, save=True, detrend_ft=True, publication=False, mad_filter=False, outdir=None):
+def plot_h5(
+        h5_file,
+        save=True,
+        detrend_ft=True,
+        publication=False,
+        mad_filter=False,
+        outdir=None,
+):
     """
     Plot the h5 candidates
 
@@ -88,26 +102,33 @@ def plot_h5(h5_file, save=True, detrend_ft=True, publication=False, mad_filter=F
         None
 
     """
[…]
