+++ your/candidate.py
@@ -27,14 +27,25 @@ class Candidate(Your):
         kill_mask (numpy.ndarray): Boolean mask of channels to kill
     """
 
-    def __init__(self, fp=None, dm=None, tcand=0, width=0, label=-1, snr=0, min_samp=256, device=0, kill_mask=None):
+    def __init__(
+            self,
+            fp=None,
+            dm=None,
+            tcand=0,
+            width=0,
+            label=-1,
+            snr=0,
+            min_samp=256,
+            device=0,
+            kill_mask=None,
+    ):
         Your.__init__(self, fp)
         self.dm = dm
         self.tcand = tcand
         self.width = width
         self.label = label
         self.snr = snr
-        self.id = f'cand_tstart_{self.tstart:.12f}_tcand_{self.tcand:.7f}_dm_{self.dm:.5f}_snr_{self.snr:.5f}'
+        self.id = f"cand_tstart_{self.tstart:.12f}_tcand_{self.tcand:.7f}_dm_{self.dm:.5f}_snr_{self.snr:.5f}"
         self.data = None
         self.dedispersed = None
         self.dmt = None
@@ -58,45 +69,55 @@ class Candidate(Your):
         """
         cand_id = self.id
         if fnout is None:
-            fnout = cand_id + '.h5'
+            fnout = cand_id + ".h5"
         if out_dir is not None:
-            if out_dir[-1] != '/':
-                out_dir = out_dir + '/'
+            if out_dir[-1] != "/":
+                out_dir = out_dir + "/"
             fnout = out_dir + fnout
-        logger.info(f'Saving h5 file {fnout}.')
-        with h5py.File(fnout, 'w') as f:
-            f.attrs['cand_id'] = cand_id
-            f.attrs['tcand'] = self.tcand
-            f.attrs['dm'] = self.dm
-            f.attrs['dm_opt'] = self.dm_opt
-            f.attrs['snr'] = self.snr
-            f.attrs['snr_opt'] = self.snr_opt
-            f.attrs['width'] = self.width
-            f.attrs['label'] = self.label
+        logger.info(f"Saving h5 file {fnout}.")
+        with h5py.File(fnout, "w") as f:
+            f.attrs["cand_id"] = cand_id
+            f.attrs["tcand"] = self.tcand
+            f.attrs["dm"] = self.dm
+            f.attrs["dm_opt"] = self.dm_opt
+            f.attrs["snr"] = self.snr
+            f.attrs["snr_opt"] = self.snr_opt
+            f.attrs["width"] = self.width
+            f.attrs["label"] = self.label
 
-            f.attrs['filelist'] = self.your_header.filelist
+            f.attrs["filelist"] = self.your_header.filelist
 
             # Copy over header information as attributes
             file_header = vars(self.your_header)
             for key in file_header.keys():
-                if key == 'dtype':
+                if key == "dtype":
                     f.attrs[key] = np.dtype(file_header[key]).name
                 else:
                     f.attrs[key] = file_header[key]
 
-            f.attrs['tsamp'] = self.your_header.tsamp
-            f.attrs['nchans'] = self.your_header.nchans
-            f.attrs['foff'] = self.your_header.foff
-            f.attrs['nspectra'] = self.your_header.nspectra
+            f.attrs["tsamp"] = self.your_header.tsamp
+            f.attrs["nchans"] = self.your_header.nchans
+            f.attrs["foff"] = self.your_header.foff
+            f.attrs["nspectra"] = self.your_header.nspectra
 
-            freq_time_dset = f.create_dataset('data_freq_time', data=self.dedispersed, dtype=self.dedispersed.dtype,
-                                              compression="gzip", compression_opts=9)
+            freq_time_dset = f.create_dataset(
+                "data_freq_time",
+                data=self.dedispersed,
[…]
