+++ your/utils/astro.py
@@ -1,7 +1,7 @@
 import numpy as np
 
-ARCSECTORAD = float('4.8481368110953599358991410235794797595635330237270e-6')
-RADTODEG = float('57.295779513082320876798154814105170332405472466564')
+ARCSECTORAD = float("4.8481368110953599358991410235794797595635330237270e-6")
+RADTODEG = float("57.295779513082320876798154814105170332405472466564")
 
 
 def dec2deg(src_dej):
@@ -12,7 +12,8 @@ def dec2deg(src_dej):
         src_dej (float): Source dec
     """
     sign = 1.0
-    if (src_dej < 0): sign = -1.0;
+    if src_dej < 0:
+        sign = -1.0
     xx = np.fabs(src_dej)
     dd = int(np.floor(xx / 10000.0))
     mm = int(np.floor((xx - dd * 10000.0) / 100.0))

[…]
