### 2020-10-15 f872ef9
Added --no_log_file option
Signed-off-by: Devansh Agarwal <devansh.kv@gmail.com>
