+++ bin/your_heimdall.py
@@ -131,43 +131,57 @@ if __name__ == "__main__":
     parser.add_argument(
         "--no_log_file", help="Do not write a log file", action="store_true"
     )
-    args = parser.parse_args()
+    values = parser.parse_args()
 
-    if args.output_dir is None:
-        args.output_dir = "{0}/{1}".format(
-            os.getcwd(), (".").join(os.path.basename(args.files[0]).split(".")[:-1])
+    if values.output_dir is None:
+        values.output_dir = "{0}/{1}".format(
+            os.getcwd(), (".").join(os.path.basename(values.files[0]).split(".")[:-1])
         )
-        os.makedirs(args.output_dir, exist_ok=True)
+        os.makedirs(values.output_dir, exist_ok=True)
 
     logging_format = (
         "%(asctime)s - %(funcName)s -%(name)s - %(levelname)s - %(message)s"
     )
 
-    if args.verbose > 0:
-        logging.basicConfig(level=logging.DEBUG, format=logging_format)
-    else:
-        logging.basicConfig(level=logging.INFO, format=logging_format)
+    log_filename = (
+            values.output_dir
+            + "/"
+            + datetime.utcnow().strftime("your_heimdall_%Y_%m_%d_%H_%M_%S_%f.log")
+    )
 
-    if args.no_log_file:
-        log_filename = (
-                args.output_dir
-                + "/"
-                + datetime.utcnow().strftime("your_heimdall_%Y_%m_%d_%H_%M_%S_%f.log")
-        )
-
-        fileHandler = logging.FileHandler(log_filename)
-        fileHandler.setFormatter(logging.Formatter(logging_format))
-        logging.getLogger().addHandler(fileHandler)
+    if not values.no_log_file:
+        if values.verbose:
+            logging.basicConfig(
+                filename=log_filename,
+                level=logging.DEBUG,
+                format=logging_format,
+            )
+        else:
+            logging.basicConfig(
+                filename=log_filename,
+                level=logging.INFO,
+                format=logging_format
+            )
     else:
-        rich_handler = RichHandler(rich_tracebacks=True)
-        logging.getLogger().addHandler(rich_handler)
+        if values.verbose:
+            logging.basicConfig(
+                level=logging.DEBUG,
+                format=logging_format,
+                handlers=[RichHandler(rich_tracebacks=True)],
+            )
+        else:
+            logging.basicConfig(
+                level=logging.INFO,
+                format=logging_format,
+                handlers=[RichHandler(rich_tracebacks=True)],
+            )
 
     logging.info("Input Arguments:-")
     for arg, value in sorted(vars(args).items()):
         logging.info("%s: %r", arg, value)
 
-    your_object = Your(file=args.files)
-    max_delay = your_object.dispersion_delay(dms=np.max(args.dm))
+    your_object = Your(file=values.files)
+    max_delay = your_object.dispersion_delay(dms=np.max(values.dm))
     dispersion_delay_samples = np.ceil(max_delay / your_object.your_header.tsamp)
     logging.info(f"Max Dispersion delay = {max_delay} s")
     logging.info(f"Max Dispersion delay = {dispersion_delay_samples} samples")
@@ -179,26 +193,26 @@ if __name__ == "__main__":
             np.max([(2 ** np.ceil(np.log2(dispersion_delay_samples))), 2 ** 18])
         )
 
-    if args.channel_end > 0:
-        c_max = args.channel_end
+    if values.channel_end > 0:
+        c_max = values.channel_end
     else:
         c_max = None
 
     your_dada_writer = Writer(
[…]
