+++ bin/your_candmaker.py
@@ -217,28 +217,37 @@ if __name__ == "__main__":
     logging_format = (
         "%(asctime)s - %(funcName)s -%(name)s - %(levelname)s - %(message)s"
     )
-
-    if values.verbose:
-        logging.basicConfig(
-            level=logging.DEBUG, format=logging_format
-        )
-    else:
-        logging.basicConfig(
-            level=logging.INFO, format=logging_format
-        )
+    log_filename = (
+            values.fout
+            + "/"
+            + datetime.utcnow().strftime("your_candmaker_%Y_%m_%d_%H_%M_%S_%f.log")
+    )
 
     if not values.no_log_file:
-        log_filename = (
-                values.fout
-                + "/"
-                + datetime.utcnow().strftime("your_candmaker_%Y_%m_%d_%H_%M_%S_%f.log")
-        )
-        fileHandler = logging.FileHandler(log_filename)
-        fileHandler.setFormatter(logging.Formatter(logging_format))
-        logging.getLogger().addHandler(fileHandler)
+        if values.verbose:
+            logging.basicConfig(
+                filename=log_filename,
+                level=logging.DEBUG,
+                format=logging_format,
+            )
+        else:
+            logging.basicConfig(
+                filename=log_filename,
+                level=logging.INFO,
+                format=logging_format
+            )
     else:
-        rich_handler = RichHandler(rich_tracebacks=True)
-        logging.getLogger().addHandler(rich_handler)
+        if values.verbose:
+            logging.basicConfig(
+                level=logging.DEBUG,
+                format=logging_format,
+                handlers=[RichHandler(rich_tracebacks=True)]
+            )
+        else:
+            logging.basicConfig(
+                level=logging.INFO,
+                format=logging_format,
+                handlers=[RichHandler(rich_tracebacks=True)])
 
     if -1 not in values.gpu_id:
         from numba.cuda.cudadrv.driver import CudaAPIError
