+++ bin/your_writer.py
@@ -75,9 +75,7 @@ if __name__ == "__main__":
         required=False,
     )
     parser.add_argument(
-        "--no_progress",
-        help="Do not show the progress bar",
-        action="store_false",
+        "--no_progress", help="Do not show the progress bar", action="store_false",
     )
     parser.add_argument(
         "-r",
@@ -133,15 +131,11 @@ if __name__ == "__main__":
     if not values.no_log_file:
         if values.verbose:
             logging.basicConfig(
-                filename=log_filename,
-                level=logging.DEBUG,
-                format=logging_format,
+                filename=log_filename, level=logging.DEBUG, format=logging_format,
             )
         else:
             logging.basicConfig(
-                filename=log_filename,
-                level=logging.INFO,
-                format=logging_format
+                filename=log_filename, level=logging.INFO, format=logging_format
             )
     else:
         if values.verbose:
