+++ formats/fitswriter/fitswriter.md
@@ -15,11 +15,17 @@ the DATA.
 
 Args: 
 
+
     outfile: path to the output fits file to write to
+
     your_object: your object with the input Filterbank file
+
     npsub: number of spectra in a subint
+
     nstart: start sample to read from (for the input file)
+
     nsamp: number of spectra to read
+
     chan_freqs: array with frequencies of all the channels
 
 
