### 2020-10-15 d9af176
fixed variable names
Signed-off-by: Devansh Agarwal <devansh.kv@gmail.com>
