### 2020-10-15 f353f39
Removed a redundant argument
Signed-off-by: Devansh Agarwal <devansh.kv@gmail.com>
