### 2020-10-15 b4dda7f
Fixed a bug with get_chunk
Signed-off-by: Devansh Agarwal <devansh.kv@gmail.com>
