### 2020-10-17 87ad6fe
fixed cpu candmaker bug
Signed-off-by: Devansh Agarwal <devansh.kv@gmail.com>
