### 2020-10-24 2747047
Removed the unused argument
Signed-off-by: Devansh Agarwal <devansh.kv@gmail.com>
