### 2020-11-01 thepetabyteproject/your#46 @pravirkr
pip install does not install required dependencies
While doing a fresh install from the git repo as suggested in docs,
`pip install git+https://github.com/thepetabyteproject/your.git`
I get the following error:
```
    Traceback (most recent call last):
      File "<string>", line 1, in <module>
      File "/tmp/pip-req-build-ee84oxd9/setup.py", line 5, in <module>
        import your
      File "/tmp/pip-req-build-ee84oxd9/your/__init__.py", line 1, in <module>
        from your.writer import *
      File "/tmp/pip-req-build-ee84oxd9/your/writer.py", line 9, in <module>
        from rich.progress import Progress
    ModuleNotFoundError: No module named 'rich'
```
The easy solution is to add more info to the installation instructions (e.g. to use the requirements.txt). A long-term fix can be to define the version either in setup.py or in a separate configuration file.

This issue is part of my review for JOSS.
