### 2020-11-01 thepetabyteproject/your#47 @pravirkr
MatplotlibDeprecationWarning
When running the plotting functions, this warning pops up:
```
MatplotlibDeprecationWarning: Support for setting the 'text.latex.preamble' or
'pgf.preamble' rcParam to a list of strings is deprecated since 3.3 and will be
removed two minor releases later; set it to a single string instead.
    self[key] = other[key]

```
It would be good to fix this warning. I suppose putting all LaTex commands in a single string will do.
This issue is part of my JOSS review.
