### 2020-11-01 thepetabyteproject/your#48 @pravirkr
Documentation Issues
Some small issues and typos (with suggestions) I found in the documentation:

- [x] update the hyperlinks (e.g., [https://github.com/devanshkv/your/issues](https://github.com/devanshkv/your/issues)). I guess this is due to the transfer of repo ownership.

### Getting Started

- [x]  The docs do not explicitly state the target audience as required for a statement of need. I will suggest adding one or two sentences in the README explaining use for a general audience. Also, hyperlinking sigproc, psrfits, etc to their proper references might be a good idea.

- [x] filterbanks -> filterbank, go -> convert

### Contributing Guide
- [x] May be highlight or italicize `your` to differentiate from your. appropritate -> appropriate
- [x] making -> make (in code of conduct)

### Reading files
- [x] the a -> a, the filterbanks -> a filterbank file
- [x] highlight `your`
- [x] fits -> psrfits, fil -> filterbank, as as -> as

### Candidates
- [x] some context re how to obtain these candidate parameters (e.g., Heimdall) would be great. Since this is also an important part of `your`, it can also be added to the statement of need in README.
- [x] candiate -> candidate, paramters -> parameters, approimate -> approximate, -SNR -> SNR, DM Time -> DM-Time
- [x] this will take data from the filterbank file -> `get_chunk()` will extract data from the source file.
- [x] frequency-time -> frequency-time data, Single pulse search pipelines lines -> Single-pulse search pipelines
- [x] `from candidate import crop` results in ModuleNotFoundError
- [x] h5 files -> HDF format file

### RFI Mitigation
- [x] mitigation -> RFI Mitigation, usefull -> useful, filterbank filterbank -> filterbank
- [x] the PSR B0329+54 single-pulse data is not part of the package. Maybe including the data file as a test case will be helpful to the users.

### Command Line Interface
- [x] fits -> psrfits, your header -> either highlight `your` or just drop it in this case.
- [x] `your_heimdall.py` description is not very clear.
- [x] I will suggest to add some more description to these cmd line tools.

### API - Your
- [x] exactly device the -> divide the. This note needs more clarification.
- [x] time_decimation_factor -> two different definitions used (decimate in time with this factor and Number of time samples to average). I like the second one. Same goes for frequency_decimation_factor
- [x] pre decimation -> pre-decimation

This issue is part of my JOSS review.
