### 2020-11-01 thepetabyteproject/your#49 @pravirkr
Paper Review
The submitted draft looks great! Few comments:

- [x] fast radio transients like pulsars -> I do not think pulsars should be classified as transients; maybe add giant pulses or single-pulse from pulsars
- [x] "streamline the pipeline" -> search pipeline?
- [x] "candidate events:" -> candidate events,
- [x] Fix citation for Astropy (missing @ in astropy:2018)

Other suggestions:

- citation for Heimdall -> I prefer B. R. Barsdell [thesis](https://ui.adsabs.harvard.edu/abs/2012PhDT.......418B) as the more appropriate citation for Heimdall than the usually cited reference in the literature (which has more to do with `dedisp`)

This issue is part of my JOSS review.
