### 2020-11-01 #2750 @pravirkr
The software package `your` is well-organized with a nice interface. The documentation is very detailed with lots of good examples. This package is going to be very useful for the FRB/radio transients science.

I have opened the following issues within the main repository:
- thepetabyteproject/your#46
- thepetabyteproject/your#47
- thepetabyteproject/your#48
- thepetabyteproject/your#49

Most of these are very minor issues. Apologies for nitpicking (reviewing for the first time).
