### 2020-11-01 ae09be5
Added all the RFI methods
Signed-off-by: Devansh Agarwal <devansh.kv@gmail.com>
