+++ requirements.txt
@@ -7,3 +7,4 @@ numba>=0.48.0
 astropy>=4.0
 pandas>=1.0.3
 rich>=8.0
+SciencePlots>=1.0.6
