+++ CONTRIBUTING.md
@@ -2,11 +2,11 @@
 
 #### **Did you find a bug?**
 
-* **Ensure the bug was not already reported** by searching on GitHub under [Issues](https://github.com/devanshkv/your/issues).
+* **Ensure the bug was not already reported** by searching on GitHub under [Issues](https://github.com/thepetabyteproject/your/issues).
 
-* If you're unable to find an open issue addressing the problem, [open a new one](https://github.com/devanshkv/your/issues/new). Be sure to include a **title and clear description**, as much relevant information as possible, and a **code sample** or an **executable test case** demonstrating the expected behavior that is not occurring.
+* If you're unable to find an open issue addressing the problem, [open a new one](https://github.com/thepetabyteproject/your/issues/new). Be sure to include a **title and clear description**, as much relevant information as possible, and a **code sample** or an **executable test case** demonstrating the expected behavior that is not occurring.
 
-* If possible, use the [bug report template](https://github.com/devanshkv/your/blob/master/.github/ISSUE_TEMPLATE/bug_report.md) to create the issue.
+* If possible, use the [bug report template](https://github.com/thepetabyteproject/your/blob/master/.github/ISSUE_TEMPLATE/bug_report.md) to create the issue.
 
 #### **Did you write a patch that fixes a bug?**
 
@@ -17,16 +17,16 @@
 
 #### **Do you intend to add a new feature or change an existing one?**
 
-* Suggest your change using the [feature request tempelate](https://github.com/devanshkv/your/blob/master/.github/ISSUE_TEMPLATE/feature_request.md) and start writing code.
+* Suggest your change using the [feature request tempelate](https://github.com/thepetabyteproject/your/blob/master/.github/ISSUE_TEMPLATE/feature_request.md) and start writing code.
 
 * Make sure you write appropritate documentation (we use Google Docstrings) and corresponding tests (using pytest).
 
 #### **Do you have questions about the source code?**
 
-* Ask any question about how to use your [here](https://github.com/devanshkv/your/issues/new).
+* Ask any question about how to use your [here](https://github.com/thepetabyteproject/your/issues/new).
 
 
-Your is a volunteer effort. We encourage you to pitch in and [join the team](https://github.com/devanshkv/your/graphs/contributors)!
+Your is a volunteer effort. We encourage you to pitch in and [join the team](https://github.com/thepetabyteproject/your/graphs/contributors)!
 
 Thanks! :heart: :heart: :heart:
 
