### 2020-11-01 thepetabyteproject/your#48 @devanshkv
@pravirkr, the code of conduct is from [contributor-covenant.org](https://www.contributor-covenant.org/version/1/4/code-of-conduct/), and we would like to refrain from making changes to it.
