+++ bin/your_writer.py
@@ -12,6 +12,7 @@ from datetime import datetime
 from rich.logging import RichHandler
 
 from your import Your
+from your.utils.misc import YourArgparseFormatter
 from your.writer import Writer
 
 logger = logging.getLogger(__name__)
@@ -20,7 +21,7 @@ if __name__ == "__main__":
     parser = argparse.ArgumentParser(
         prog="your_writer.py",
         description="Convert/Write files from any format to a single file in any format.",
-        formatter_class=argparse.ArgumentDefaultsHelpFormatter,
+        formatter_class=YourArgparseFormatter,
     )
     parser.add_argument("-v", "--verbose", help="Be verbose", action="store_true")
     parser.add_argument(
