+++ paper/paper.bib
@@ -253,3 +253,14 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2020MNRAS.497.1661A},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@PHDTHESIS{barsdell2012heimdall,
+       author = {{Barsdell}, B.~R.},
+        title = "{Advanced architectures for astrophysical supercomputing}",
+       school = {Swinburne University of Technology},
+         year = 2012,
+        month = jan,
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2012PhDT.......418B},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
