### 2020-11-02 thepetabyteproject/your#49 @KshitijAggarwal
Made the suggested changed with  24e2e15

For "fast radio transients like pulsars": I guess it is a personal preference. I think of them as fast transients which are periodic. Each pulse would still be a transient. I've modified it to ".. like pulsar single pulses, ..."

For "candidate events:" as we are listing some analysis that can be done on the candidate events, I think a colon is better here.
