### 2020-11-02 380f6bd
added more description in readme
