### 2020-11-02 a59bd05
added more description to command line scripts
