+++ bin/your_h5plotter.py
@@ -13,6 +13,7 @@ from rich.progress import Progress
 
 from your.utils.misc import YourArgparseFormatter
 from your.utils.plotter import plot_h5
+import textwrap
 
 os.environ["HDF5_USE_FILE_LOCKING"] = "FALSE"
 matplotlib.use("Agg")
@@ -33,8 +34,15 @@ def mapper(save, detrend_ft, publication, mad_filter, out_dir, h5_file):
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(
         prog="your_h5plotter.py",
-        description="Plot candidate h5 files",
+        description="Plot candidate h5 files.",
         formatter_class=YourArgparseFormatter,
+        epilog=textwrap.dedent(
+            """\
+            This script can be used to plot the candidate HDF5 format files generated by your_candmaker.py 
+            and the ones generated by the Candidate class and save them as a png file. 
+            """
+        )
+
     )
     parser.add_argument("-v", "--verbose", help="Be verbose", action="store_true")
     parser.add_argument(
