+++ bin/your_combine_mocks.py
@@ -509,7 +509,7 @@ def all_files(direct, nstart, nsamp, outdir):
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(
         prog="your_combine_mocks.py",
-        description="Combine two bands from mock spectrometer to a filterbank file.",
+        description="Combine two bands generated by mock spectrometer (at Arecibo Telescope) to a single filterbank file.",
         formatter_class=YourArgparseFormatter,
     )
     parser.add_argument("-v", "--verbose", help="Be verbose", action="store_true")
