### 2020-11-02 c8c04ac
minor changes in docs
