+++ your/your.py
@@ -171,12 +171,14 @@ class Your(PsrfitsFile, SigprocFile):
         Args:
             nstart (int): start sample
             nsamp (int): number of samples to read
-            time_decimation_factor (int): decimate in time with this factor
-            frequency_decimation_factor (int): decimate in frequency with this factor
+            time_decimation_factor (int): number of time samples to average
+            frequency_decimation_factor (int): number of frequency channels to average
             pol (int): which polarization to chose
 
         Note:
-            Both decimation factors should exactly device the nsamp or nchans
+            The decimation (both in time and frequency) is done on the data read i.e containing `nsamp` number of samples
+            and `nchans` number of channels. Therefore, both decimation factors should exactly divide the
+            nsamps or nchans respectively.
 
         Returns:
             numpy.ndarray: 2D numpy array of data
@@ -301,10 +303,10 @@ class Header:
         center_freq (float): Center frequency of the data.
         time_decimation_factor (int): Number of time samples to average
         frequency_decimation_factor (int): Number of frequency channels to average
-        native_tsamp (float): Sampling time of the data pre decimation (seconds)
-        native_foff (float): Channel bandwidth of the data pre decimation (MHz)
-        native_nchans : Number of channels in the data pre decimation
-        native_nspectra: Number of spectra in the data pre decimation
+        native_tsamp (float): Sampling time of the data pre-decimation (seconds)
+        native_foff (float): Channel bandwidth of the data pre-decimation (MHz)
+        native_nchans : Number of channels in the data pre-decimation
+        native_nspectra: Number of spectra in the data pre-decimation
         dtype: dtype of the (read) data
         nbits (int): Number of bits in the data
         tstart (float): Start MJD of the data
