### 2020-11-06 3a5f57f
removing SciencePlots
