### 2020-11-06 97e3f86
reallly getting rid of SciencePlots
