### 2020-11-06 a3f367b
added install instructions for git clone
