### 2020-11-06 7e4c5f3
small changes in docs
