+++ bin/your_h5plotter.py
@@ -38,8 +38,7 @@ if __name__ == "__main__":
         formatter_class=YourArgparseFormatter,
         epilog=textwrap.dedent(
             """\
-            This script can be used to plot the candidate HDF5 format files generated by your_candmaker.py 
-            and the ones generated by the Candidate class and save them as a png file. 
+            This script can be used to plot and save the candidate HDF5 format files generated by your_candmaker.py or by the Candidate class. 
             """
         )
 
