+++ bin/your_h5plotter/your_h5plotter.md
@@ -5,12 +5,22 @@ your_h5plotter.py
 # Description
 
 
-Plot candidate h5 files
+Plot candidate h5 files.
+# Epilog
+
+
+
+This script can be used to plot and save the candidate HDF5 format files generated by your_candmaker.py or by the Candidate class. 
+
+
 # Usage:
 
 
 ```bash
-usage: your_h5plotter.py [-h] [-v] [-f FILES [FILES ...]] [-c RESULTS_CSV] [--publish] [--no_detrend_ft] [--no_save] [-o OUT_DIR] [-mad [MAD_FILTER]] [-n NPROC] [--no_progress] [--no_log_file]
+usage: your_h5plotter.py [-h] [-v] [-f FILES [FILES ...]] [-c RESULTS_CSV]
+                         [--publish] [--no_detrend_ft] [--no_save]
+                         [-o OUT_DIR] [-mad [MAD_FILTER]] [-n NPROC]
+                         [--no_progress] [--no_log_file]
 
 ```
 # Arguments
