+++ bin/your_combine_mocks/your_combine_mocks.md
@@ -5,12 +5,14 @@ your_combine_mocks.py
 # Description
 
 
-Combine two bands from mock spectrometer to a filterbank file.
+Combine two bands generated by mock spectrometer (at Arecibo Telescope) to a single filterbank file.
 # Usage:
 
 
 ```bash
-usage: your_combine_mocks.py [-h] [-v] [-f1 FIRST_BAND] [-f2 SECOND_BAND] [-s NSTART] [-n NSAMP] [-o OUTDIR] [-fil FIL_NAME] [-a ALL_FILES] [--no_log_file]
+usage: your_combine_mocks.py [-h] [-v] [-f1 FIRST_BAND] [-f2 SECOND_BAND]
+                             [-s NSTART] [-n NSAMP] [-o OUTDIR]
+                             [-fil FIL_NAME] [-a ALL_FILES] [--no_log_file]
 
 ```
 # Arguments
