+++ bin/your_rfimask/your_rfimask.md
@@ -10,7 +10,11 @@ Make Bad channel mask
 
 
 ```bash
-usage: your_rfimask.py [-h] [-f FILES [FILES ...]] [-sg] [-frequency_window FILTER_WINDOW [FILTER_WINDOW ...]] [-sigma SIGMA [SIGMA ...]] [-o OUTPUT_DIR] [--no_log_file]
+usage: your_rfimask.py [-h] [-f FILES [FILES ...]] [-n NSPECTRA]
+                       [-sk_sigma SPECTRAL_KURTOSIS_SIGMA]
+                       [-sg_sigma SAVGOL_SIGMA]
+                       [-sg_frequency SAVGOL_FREQUENCY_WINDOW] [-o OUTPUT_DIR]
+                       [--no_log_file] [-v]
 
 ```
 # Arguments
@@ -19,8 +23,10 @@ usage: your_rfimask.py [-h] [-f FILES [FILES ...]] [-sg] [-frequency_window FILT
 | :---: | :---: | :---: | :---: |
 |`-h`|`--help`||show this help message and exit|
 |`-f`|`--files`|`None`|filterbank or psrfits|
-|`-sg`|`--apply_savgol`||Apply savgol filter to zap bad channels|
-|`-frequency_window`|`--filter_window`|`[15]`|Window size (MHz) for savgol filter|
-|`-sigma`|`--sigma`|`[6]`|Sigma for the savgol filter|
-|`-o`|`--output_dir`|`.`|Output dir for heimdall candidates|
+|`-n`|`--nspectra`|`8192`|Number of spectra to read and apply filters to|
+|`-sk_sigma`|`--spectral_kurtosis_sigma`|`0`|Sigma for spectral kurtosis based RFI mitigation, if set to 0 this method will not be used.|
+|`-sg_sigma`|`--savgol_sigma`|`0`|Sigma for Savgol filter for RFI mitigation, if set to 0 this method will not be used.|
+|`-sg_frequency`|`--savgol_frequency_window`|`15`|Filter window for savgol filter (in MHz), only applied if -rfi_flag is used.|
+|`-o`|`--output_dir`|`.`|Output dir for saving the mask|
 ||`--no_log_file`||Do not write a log file|
+|`-v`|`--verbose`||Be verbose|
