### 2020-11-09 #2750 @paulray
I'm supposed to answer this question: "Does the full list of paper authors seem appropriate and complete?"
I don't know how to judge that, other than from git commits. Is there a statement of the various author contributions somewhere? I didn't see it in the paper.
