### 2020-11-09 #2750 @KshitijAggarwal
Hi @paulray, I can answer that here:
@KshitijAggarwal  @devanshkv: equal contribution in terms of code, ideas, discussion, structure and documentation
@scottransom: original author of [psrfits](https://github.com/scottransom/presto/blob/master/python/presto/psrfits.py), which was modified for `your`
@demorest: original author of [pysigproc](https://github.com/demorest/pysigproc/blob/master/pysigproc.py), which was modified for `your`
@ReshmaAnnaThomas and @wcfiore: wrote + contributed to `filwriter.py`
@josephwkania: worked on RFI mitigation functions, `your_viewer`
@xiggystardust, Duncan Lorimer, Maura McLaughlin, @nategarver-daniels: contributed through discussions and structure
@rwharton: original author of [fitswriter](https://github.com/rwharton/fil2psrfits), which was then modified for `your`
