### 2020-11-09 #2750 @paulray
Regarding my review, I want to disclose the following potential COIs.

- Scott Ransom and I are longtime collaborators. We are co-Is on several funded guest observer proposals and work closely together on the open source PINT project and many aspects of Fermi and pulsar science.
- Paul Demorest is a collaborator of mine on several projects.
- Maura McLaughlin is a close collaborator on Fermi pulsar searches and some other projects.
- All of the above people as well as Kshitij Aggarwal, Sarah Burke-Spolaor, Dunc Lorimer, Nate Garver-Daniels, and William Fiore are (or were) members of the NANOGrav collaboration that I am a full member of. We all collaborate and co-author many papers together.

I just now realized that the submitting author is a NANOGrav member.  I don't think that is a serious conflict since I wasn't even aware of it. According to the statement above, my other potential COIs generally have a largely advisory role in this project or have prior code that was subsumed into this project.

I don't think these potential COIs affect my ability to make an impartial review, so I propose that they be waived.
