### 2020-11-09 #2750 @dfm
@paulray: Thanks for bringing these up! I agree that these potential conflicts seem ok to be waived.

@KshitijAggarwal: do you or co-authors don't have any concerns about this? Please feel free to let me know here or offline via email if you'd prefer. Thanks!
