### 2020-11-09 #2750 @paulray
I enjoyed reviewing the paper and the code and trying it out on same data I had. I found it well explained, easy to install and the documentation was very clear. Nice work!
