### 2020-11-10 #2750 @KshitijAggarwal
@pravirkr @dfm, we are submitting an NSF proposal for a project (the petabyte project) which is going to rely on this library. The NSF proposal is due pretty soon. I also noticed that only a few checklist items are left in the review. So I wanted to ask if it would be possible to complete this review in the next day or so? This way we can properly cite it in our proposal, and also upload it to arxiv.
If not, is it okay if we upload the paper to arxiv before the review process is complete?
