### 2020-11-10 #2750 @dfm
You're welcome to post the pre-print to ArXiv, but sometimes ArXiv doesn't accept JOSS papers so no guarantees that it will work!

@pravirkr: you definitely shouldn't feel pressured to finish your review faster than planned!
