### 2020-11-11 #2750 @pravirkr
@dfm: Thanks. I was waiting for the resolution of the above issues.
@KshitijAggarwal: Thanks for addressing the issues. Looks good now! The only thing left to do IMO is to release the updated version (0.5.9, I guess).

I also recommend acceptance of this work.
