### 2020-11-12 #2750 @dfm
@KshitijAggarwal: I've just submitted a small pull request with some formatting changes. After you take a look at that, here are the steps to run:

1. Comment `@whedon generate pdf` on this thread and carefully read through the manuscript to make sure that you're happy with it because it can be hard to update later. You should especially double check author names and affiliations.
2. Update the version number on the software and create an archive of that release (using Zenodo or similar) and report the archive DOI and new version number to this thread. **NOTE: The title and author list in the archive metadata must exactly match this manuscript** (you can generally update the metadata even after the archive has been made).

Let me know if you have any issues or questions!
