### 2020-11-12 ea1b64a
Writing out journal names in full
