+++ paper/paper.bib
@@ -5,7 +5,7 @@
          {Armour}, W. and {MacMahon}, D.~H.~E. and {Siemion}, A.~P.~V. and
          {Foster}, G.},
         title = "{Initial results from a real-time FRB search with the GBT}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the Royal Astronomical Society},
      keywords = {surveys, pulsars: general, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2020,
         month = jul,
@@ -24,7 +24,7 @@ archivePrefix = {arXiv},
 @ARTICLE{nita2010,
        author = {{Nita}, G.~M. and {Gary}, D.~E.},
         title = "{The generalized spectral kurtosis estimator}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the Royal Astronomical Society},
      keywords = {instrumentation: spectrographs, methods: statistical, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2010,
         month = jul,
@@ -67,7 +67,7 @@ archivePrefix = {arXiv},
          {Perktold}, Josef and {Cimrman}, Robert and {Henriksen}, Ian and
          {Quintero}, E.~A. and {Harris}, Charles R and {Archibald}, Anne M.
          and {Ribeiro}, Ant{\^o}nio H. and {Pedregosa}, Fabian and
-         {van Mulbregt}, Paul and {Contributors}, SciPy 1. 0},
+         {van Mulbregt}, Paul},
         title = "{SciPy 1.0: Fundamental Algorithms for Scientific
                   Computing in Python}",
       journal = {Nature Methods},
@@ -108,7 +108,7 @@ Author = {{Astropy Collaboration} and {Robitaille}, T.~P. and {Tollerud}, E.~J. 
 Doi = {10.1051/0004-6361/201322068},
 Eid = {A33},
 Eprint = {1307.6212},
-Journal = {\aap},
+Journal = {Astronomy \& Astrophysics},
 Keywords = {methods: data analysis, methods: miscellaneous, virtual observatory tools},
 Month = oct,
 Pages = {A33},
@@ -124,7 +124,7 @@ Adsurl = {https://ui.adsabs.harvard.edu/#abs/2018AJ....156..123T},
[…]
