### 2020-11-12 07f64e3
Handling accents
