### 2020-11-12 #2750 @dfm
@KshitijAggarwal: I found a few more remaining formatting issues with the manuscript. Take a look at that PR and then we should be good to go!
